requirements.txt
alembic==0.9.7
asn1crypto==0.24.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
boto3==1.5.27
botocore==1.8.41
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
cognitive-face==1.3.1
cryptography==2.1.4
docutils==0.14
dominate==2.3.1
enum34==1.1.6
Flask==0.12.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
futures==3.2.0
http==0.2
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.0
opencv-python==3.4.0.12
pycparser==2.18
PyMySQL==0.8.0
python-dateutil==2.6.1
python-editor==1.0.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
SQLAlchemy==1.2.2
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
